requirements.txt
opencv-python==4.8.1.78
Pillow==10.1.0
reportlab==4.0.7
